Avoid CTRL-\ CTRL-O in insert paste
CTRL-\ CTRL-O moves the cursor like in CTRL-O at the end of the line if
the cursor is on a line with a syntax error (when using e.g. w0rp/ale).

Not using CTRL-\ CTRL-O is also nicer because with :set list, it does
not show a momentary blinking trailing whitespace.